refactor(app-viewer): reorganize file names and locations
- change file names to more accuratly match its usage
- relocate sections folder to pages
