+++ apps/ui/src/pages/applications/sections/Introduction.tsx
