+++ apps/ui/src/pages/applications/sections/SectionA.tsx
